chore: clear clippy -D warnings debt in audio/dat/proto/nav
Mechanical lint cleanup (no behavior change):
- ffxi-audio: range-index loops in the channel-major decode + interleave
  paths read clearer than iterators; module-level allow with rationale.
- ffxi-dat/ffxi-proto: rewrite needless_range_loops as iterators; simplify
  `a || (!a && b)`; drop a useless format! in build codegen. Allow the
  doc-list lints crate-wide (hand-aligned byte-layout docs are deliberate).
- ffxi-nav: emit an allow on the generated zonelines table (rounded data
  literals trip excessive_precision / approx_constant).
- plus cargo clippy --fix machine-applicable fixes.

+++ ffxi-dat/src/zone_dat.rs
@@ -85,7 +85,7 @@ mod tests {
     fn high_zone_branch_is_reachable() {
         let any_high = ZONE_DAT_TABLE
             .iter()
-            .any(|(z, f)| *z >= 256 && *f as u32 == *z as u32 + 83635);
+            .any(|(z, f)| *z >= 256 && *f == *z as u32 + 83635);
         assert!(
             any_high,
             "no zone_id >= 256 found applying the high-branch formula"
